pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pii-airlock"
version = "0.1.0"
description = "PII protection middleware for public LLM APIs - Make Public LLMs Private"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "PII-AIRLOCK Team" }
]
keywords = ["pii", "anonymization", "llm", "privacy", "presidio", "proxy"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]

dependencies = [
    "presidio-analyzer>=2.2.0,<3.0.0",
    "presidio-anonymizer>=2.2.0,<3.0.0",
    "spacy>=3.7.0,<4.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "fastapi>=0.100.0,<1.0.0",
    "uvicorn[standard]>=0.23.0,<1.0.0",
    "httpx>=0.24.0,<1.0.0",
    "redis>=4.5.0,<6.0.0",
    "pyyaml>=6.0,<7.0.0",
    "python-json-logger>=2.0.0,<3.0.0",
    "prometheus-client>=0.17.0,<1.0.0",
    "slowapi>=0.1.8,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/pii-airlock/pii-airlock"
Documentation = "https://github.com/pii-airlock/pii-airlock#readme"
Repository = "https://github.com/pii-airlock/pii-airlock"

[tool.hatch.build.targets.wheel]
packages = ["src/pii_airlock"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --cov=pii_airlock --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "W", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true
